backend/requirements/reqs.txt
fastapi==0.109.0
uvicorn[standard]==0.30.1
python-keycloak==4.2.2
ipython==7.34.0
